Linux

+++ src/futil/unittest/futil_test.cc
@@ -12,18 +12,23 @@ class tmock_test
 
     TMOCK_TEST(test_path_count_components)
     {
-        tmock::assert_equiv(futil::path("////a///b/c///").count_components(),3);
-        tmock::assert_equiv(futil::path("aaa/bb/cccc").count_components(),3);
-        tmock::assert_equiv(futil::path("aaa/bb/").count_components(),2);
-        tmock::assert_equiv(futil::path("/aaa/bb/").count_components(),2);
-        tmock::assert_equiv(futil::path("/aaa/bb").count_components(),2);
-        tmock::assert_equiv(futil::path("/aaa/b").count_components(),2);
-        tmock::assert_equiv(futil::path("/a/").count_components(),1);
-        tmock::assert_equiv(futil::path("/aa/").count_components(),1);
-        tmock::assert_equiv(futil::path("a/").count_components(),1);
-        tmock::assert_equiv(futil::path("aa/").count_components(),1);
-        tmock::assert_equiv(futil::path("aa").count_components(),1);
-        tmock::assert_equiv(futil::path("a").count_components(),1);
+        tmock::assert_equiv(futil::path("////a///b/c///").count_components(),
+                            (size_t)3);
+        tmock::assert_equiv(futil::path("aaa/bb/cccc").count_components(),
+                            (size_t)3);
+        tmock::assert_equiv(futil::path("aaa/bb/").count_components(),
+                            (size_t)2);
+        tmock::assert_equiv(futil::path("/aaa/bb/").count_components(),
+                            (size_t)2);
+        tmock::assert_equiv(futil::path("/aaa/bb").count_components(),
+                            (size_t)2);
+        tmock::assert_equiv(futil::path("/aaa/b").count_components(),(size_t)2);
+        tmock::assert_equiv(futil::path("/a/").count_components(),(size_t)1);
+        tmock::assert_equiv(futil::path("/aa/").count_components(),(size_t)1);
+        tmock::assert_equiv(futil::path("a/").count_components(),(size_t)1);
+        tmock::assert_equiv(futil::path("aa/").count_components(),(size_t)1);
+        tmock::assert_equiv(futil::path("aa").count_components(),(size_t)1);
+        tmock::assert_equiv(futil::path("a").count_components(),(size_t)1);
     }
 
     TMOCK_TEST(test_path_join)
@@ -50,21 +55,21 @@ class tmock_test
     {
         auto p = futil::path("aaa/bb/cccc");
         auto v = p.decompose();
-        tmock::assert_equiv(v.size(),3);
+        tmock::assert_equiv(v.size(),(size_t)3);
         tmock::assert_equiv(v[0],"aaa");
         tmock::assert_equiv(v[1],"bb");
         tmock::assert_equiv(v[2],"cccc");
 
         p = futil::path("/aaa/bb/");
         v = p.decompose();
-        tmock::assert_equiv(v.size(),3);
+        tmock::assert_equiv(v.size(),(size_t)3);
         tmock::assert_equiv(v[0],"/");
         tmock::assert_equiv(v[1],"aaa");
         tmock::assert_equiv(v[2],"bb");
 
         p = futil::path("////a///b/c///");
         v = p.decompose();
-        tmock::assert_equiv(v.size(),4);
+        tmock::assert_equiv(v.size(),(size_t)4);
         tmock::assert_equiv(v[0],"/");
         tmock::assert_equiv(v[1],"a");
         tmock::assert_equiv(v[2],"b");
